requirements.txt
numpy==1.26.0
protobuf==3.20.0
opencv-python
scipy==1.15.3
fvcore==0.1.5.post20221221
segment-anything @ git+https://github.com/facebookresearch/segment-anything.git@dca509fe793f601edb92606367a655c15ac00fdf
tensorboard==2.8.0
tensorboardX==2.1
tensorflow==2.8.0
tqdm==4.67.1
timm==1.0.15
transforms3d==0.4.2
xformers==0.0.28
open3d==0.19.0
ultralytics==8.0.135
distinctipy
hydra-colorlog
hydra-core
gdown
pandas
imageio==2.37.0
pyrender
gorilla-core==0.2.7.8
gpustat==1.0.0
imgaug==0.4.0
einops==0.8.1
ruamel.yaml
pytorch-lightning
